requirements.txt
Package                   Version
------------------------- -----------
aiohappyeyeballs          2.4.0
aiohttp                   3.10.5
aiosignal                 1.3.1
altair                    5.4.1
annotated-types           0.7.0
anyio                     4.4.0
attrs                     24.2.0
blinker                   1.8.2
cachetools                5.5.0
certifi                   2024.7.4
charset-normalizer        3.3.2
click                     8.1.7
dataclasses-json          0.6.7
distro                    1.9.0
frozenlist                1.4.1
gitdb                     4.0.11
GitPython                 3.1.43
h11                       0.14.0
httpcore                  1.0.5
httpx                     0.27.2
idna                      3.8
Jinja2                    3.1.4
jiter                     0.5.0
jsonpatch                 1.33
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2023.12.1
langchain                 0.2.15
langchain-community       0.2.14
langchain-core            0.2.36
langchain-text-splitters  0.2.2
langsmith                 0.1.106
markdown-it-py            3.0.0
MarkupSafe                2.1.5
marshmallow               3.22.0
mdurl                     0.1.2
multidict                 6.0.5
mypy-extensions           1.0.0
narwhals                  1.5.5
numpy                     1.26.4
openai                    1.42.0
orjson                    3.10.7
packaging                 24.1
pandas                    2.2.2
pillow                    10.4.0
pip                       24.2
protobuf                  5.28.0
pyarrow                   17.0.0
pydantic                  2.8.2
pydantic_core             2.20.1
pydeck                    0.9.1
Pygments                  2.18.0
python-dateutil           2.9.0.post0
python-dotenv             1.0.1
pytz                      2024.1
PyYAML                    6.0.2
referencing               0.35.1
requests                  2.32.3
rich                      13.8.0
rpds-py                   0.20.0
setuptools                65.5.0
six                       1.16.0
smmap                     5.0.1
sniffio                   1.3.1
SQLAlchemy                2.0.32
streamlit                 1.38.0
tenacity                  8.5.0
toml                      0.10.2
tornado                   6.4.1
tqdm                      4.66.5
typing_extensions         4.12.2
typing-inspect            0.9.0
tzdata                    2024.1
urllib3                   2.2.2
yarl                      1.9.4
